Refactored End2End test Moved startup of app to CopyProg class
